+++ src/Symfony/Component/Validator/Tests/Fixtures/PropertyInfoLoaderEntity.php
@@ -46,4 +46,6 @@ class PropertyInfoLoaderEntity
      * })
      */
     public $alreadyPartiallyMappedCollection;
+
+    public $readOnly;
 }