+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * [BC BREAK] `SendersLocatorInterface` has an additional method:
    `getSenderByAlias()`.
+ * Removed argument `?bool &$handle = false` from `SendersLocatorInterface::getSenders`
  * A new `ListableReceiverInterface` was added, which a receiver
    can implement (when applicable) to enable listing and fetching
    individual messages by id (used in the new "Failed Messages" commands).